refactor: 캐러셀 반복 로직 분리
