+++ src/app/home/components/MovieCarousel/MovieCarousel.css.ts
