+++ src/app/home/components/MovieCarousel/movie/BoxOfficeCarousel.tsx
@@ -1,0 +1,14 @@
+import * as styles from '@/app/home/components/MovieCarousel/MovieCarousel.css'
+import MovieCarousel from '../MovieCarousel'
+
+const BoxOfficeCarousel = () => {
+  return (
+    <MovieCarousel
+      type="boxoffice"
+      className={styles.container}
+      slideContainerClassName={styles.slideContainer}
+    />
+  )
+}
+
+export default BoxOfficeCarousel
