+++ src/app/home/components/MovieCarousel/movie/DdayCarousel.tsx
@@ -1,0 +1,14 @@
+import * as styles from '@/app/home/components/MovieCarousel/MovieCarousel.css'
+import MovieCarousel from '../MovieCarousel'
+
+const DdayCarousel = () => {
+  return (
+    <MovieCarousel
+      type="dDay"
+      className={styles.container}
+      slideContainerClassName={styles.slideContainer}
+    />
+  )
+}
+
+export default DdayCarousel
