+++ src/app/home/components/MovieCarousel/movie/SeriesCarousel.tsx
@@ -1,0 +1,14 @@
+import * as styles from '@/app/home/components/MovieCarousel/MovieCarousel.css'
+import MovieCarousel from '../MovieCarousel'
+
+const SeriesCarousel = () => {
+  return (
+    <MovieCarousel
+      type="series"
+      className={styles.container}
+      slideContainerClassName={styles.slideContainer}
+    />
+  )
+}
+
+export default SeriesCarousel
