+++ src/app/home/components/MovieCarousel/movie/TopRankingCarousel.tsx
@@ -1,0 +1,14 @@
+import * as styles from '@/app/home/components/MovieCarousel/MovieCarousel.css'
+import MovieCarousel from '../MovieCarousel'
+
+const TopRankingCarousel = () => {
+  return (
+    <MovieCarousel
+      type="rank"
+      className={styles.container}
+      slideContainerClassName={styles.slideContainer}
+    />
+  )
+}
+
+export default TopRankingCarousel
